+++ services/src/main/java/org/keycloak/timer/basic/BasicTimerProviderFactory.java
@@ -18,8 +18,10 @@
 package org.keycloak.timer.basic;
 
 import org.keycloak.Config;
+import org.keycloak.cluster.ClusterProvider;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.KeycloakSessionFactory;
+import org.keycloak.storage.managers.UserStorageSyncManager;
 import org.keycloak.timer.TimerProvider;
 import org.keycloak.timer.TimerProviderFactory;
 
@@ -37,6 +39,7 @@ public class BasicTimerProviderFactory implements TimerProviderFactory {
     private int transactionTimeout;
 
     public static final String TRANSACTION_TIMEOUT = "transactionTimeout";
+    public static final String CANCEL_TASK = "cancelTask";
 
     private ConcurrentMap<String, TimerTaskContextImpl> scheduledTasks = new ConcurrentHashMap<>();
 
@@ -53,7 +56,10 @@ public class BasicTimerProviderFactory implements TimerProviderFactory {
 
     @Override
     public void postInit(KeycloakSessionFactory factory) {
-
+        try (KeycloakSession session = factory.create()) {
+            ClusterProvider clusterProvider = session.getProvider(ClusterProvider.class);
+            clusterProvider.registerListener(CANCEL_TASK, new TaskCancellationListener(factory));
+        }
     }
 
     @Override