+++ services/src/main/java/org/keycloak/timer/basic/TaskCancellationEvent.java
@@ -1,0 +1,15 @@
+package org.keycloak.timer.basic;
+
+import org.keycloak.cluster.ClusterEvent;
+
+public class TaskCancellationEvent implements ClusterEvent {
+    private final String taskName;
+
+    public TaskCancellationEvent(String taskName) {
+        this.taskName = taskName;
+    }
+
+    public String getTaskName() {
+        return taskName;
+    }
+}